+++ packages/miniapp-render/src/createConfig/page.js
@@ -1,12 +1,12 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { isMiniApp, isWeChatMiniProgram } from 'universal-env';
+
 import cache from '../utils/cache';
 import injectLifeCycle from '../bridge/injectLifeCycle';
 import createEventProxy from '../bridge/createEventProxy';
-import perf from '../utils/perf';
 import createDocument from '../document';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { isMiniApp } from 'universal-env';
 import { BODY_NODE_ID } from '../constants';
-import { createWindow } from '../window';
+import createWindow from '../window';
 
 export function getBaseLifeCycles(route, init, packageName = '') {
   return {
@@ -40,6 +40,9 @@ export function getBaseLifeCycles(route, init, packageName = '') {
       }
       // Bind page internal to page document
       this.document._internal = this;
+      if (isWeChatMiniProgram) {
+        cache.setPageInstance(this);
+      }
       this.query = query;
       // Update location page options
       this.window.history.location.__updatePageOption(query);