+++ packages/rax-miniapp-config-webpack-plugin/src/transformNativeConfig.js
@@ -1,7 +1,7 @@
 const { join } = require('path');
+const { constants: { MINIAPP, WECHAT_MINIPROGRAM, BYTEDANCE_MICROAPP } } = require('miniapp-builder-shared');
 
 const safeWriteFile = require('./safeWriteFile');
-const { MINIAPP, WECHAT_MINIPROGRAM, BYTEDANCE_MICROAPP } = require('./constants');
 
 const fileNameMap = {
   [MINIAPP]: 'mini.project.json',