+++ packages/rax-miniapp-runtime-webpack-plugin/src/generators/page.js
@@ -1,12 +1,11 @@
-const ejs = require('ejs');
 const { join } = require('path');
-const adapter = require('../adapter');
+const { platformMap } = require('miniapp-builder-shared');
+
 const getAssetPath = require('../utils/getAssetPath');
 const getSepProcessedPath = require('../utils/getSepProcessedPath');
 const addFileToCompilation = require('../utils/addFileToCompilation');
-const getTemplate = require('../utils/getTemplate');
 const { pathHelper: { getBundlePath }} = require('miniapp-builder-shared');
-const { MINIAPP } = require('../constants');
+const { RECURSIVE_TEMPLATE_TYPE } = require('../constants');
 
 function generatePageCSS(
   compilation,
@@ -15,8 +14,8 @@ function generatePageCSS(
   { target, command }
 ) {
   let pageCssContent = '/* required by usingComponents */\n';
-  const pageCssPath = `${pageRoute}.${adapter[target].css}`;
-  const subAppCssPath = `${getBundlePath(subAppRoot)}.css.${adapter[target].css}`;
+  const pageCssPath = `${pageRoute}${platformMap[target].extension.css}`;
+  const subAppCssPath = `${getBundlePath(subAppRoot)}.css${platformMap[target].extension.css}`;
   if (compilation.assets[subAppCssPath]) {
     pageCssContent += `@import "${getAssetPath(subAppCssPath, pageCssPath)}";`;
   }
@@ -34,27 +33,21 @@ function generatePageJS(
   compilation,
   pageRoute,
   pagePath,
-  nativeLifeCycles = {},
+  nativeLifeCyclesMap = {},
   commonPageJSFilePaths = [],
   subAppRoot = '',
   { target, command, pluginDir, outputPath }
 ) {
-  const pageJsContent = ejs.render(getTemplate(pluginDir, 'page.js'), {
-    render_path: `${getAssetPath(join(outputPath, 'render.js'), join(outputPath, `${pageRoute}.js`))}`,
-    route: getSepProcessedPath(pagePath),
-    native_lifecycles: `[${Object.keys(nativeLifeCycles).reduce((total, current, index) =>
-      index === 0 ? `${total}'${current}'` : `${total},'${current}'`, '')}]`,
-    init: `function init(window, document) {${commonPageJSFilePaths
-      .map(
-        filePath =>
-          `require('${getAssetPath(
-            filePath,
-            pageRoute
-          )}')(window, document)`
-      )
-      .join(';')}}`,
-    root: subAppRoot
-  });
+  const renderPath = getAssetPath('render', pageRoute);
+  const route = getSepProcessedPath(pagePath);
+  const nativeLifeCycles = `[${Object.keys(nativeLifeCyclesMap).reduce((total, current, index) => index === 0 ? `${total}'${current}'` : `${total},'${current}'`, '')}]`;
+  const init = `
+function init(window, document) {${commonPageJSFilePaths.map(filePath => `require('${getAssetPath(filePath, pageRoute)}')(window, document)`).join(';')}}`;
+
+  const pageJsContent = `
+const render = require('${renderPath}');
+${init}
+Page(render.createPageConfig('${route}', ${nativeLifeCycles}, init, '${subAppRoot}'))`;
 
   addFileToCompilation(compilation, {
     filename: `${pageRoute}.js`,
@@ -71,17 +64,17 @@ function generatePageXML(
   { target, command, outputPath }
 ) {
   let pageXmlContent;
-  if (target === MINIAPP && useComponent) {
+  if (RECURSIVE_TEMPLATE_TYPE.has(target) && useComponent) {
     pageXmlContent = '<element r="{{root}}"  />';
   } else {
-    const rootTmplFileName = 'root.' + adapter[target].xml;
-    const pageTmplFilePath = `${pageRoute}.` + adapter[target].xml;
+    const rootTmplFileName = `root${platformMap[target].extension.xml}`;
+    const pageTmplFilePath = `${pageRoute}${platformMap[target].extension.xml}`;
     pageXmlContent = `<import src="${getAssetPath(join(outputPath, rootTmplFileName), join(outputPath, pageTmplFilePath))}"/>
-    <template is="element" data="{{r: root}}"  />`;
+<template is="RAX_TMPL_ROOT_CONTAINER" data="{{r: root}}"  />`;
   }
 
   addFileToCompilation(compilation, {
-    filename: `${pageRoute}.${adapter[target].xml}`,
+    filename: `${pageRoute}${platformMap[target].extension.xml}`,
     content: pageXmlContent,
     target,
     command,
@@ -92,19 +85,24 @@ function generatePageJSON(
   compilation,
   pageConfig,
   useComponent,
+  usingComponents, usingPlugins,
   pageRoute,
-  { target, command, outputPath }
+  { target, command }
 ) {
   if (!pageConfig.usingComponents) {
     pageConfig.usingComponents = {};
   }
-  const elementPath = getAssetPath(
-    join(outputPath, 'comp'),
-    join(outputPath, `${pageRoute}.js`)
-  );
-  if (useComponent || target !== MINIAPP) {
-    pageConfig.usingComponents.element = elementPath;
+
+  if (useComponent || !RECURSIVE_TEMPLATE_TYPE.has(target)) {
+    pageConfig.usingComponents.element = getAssetPath('comp', pageRoute);
   }
+
+  Object.keys(usingComponents).forEach(component => {
+    pageConfig.usingComponents[component] = getAssetPath(usingComponents[component].path, pageRoute);
+  });
+  Object.keys(usingPlugins).forEach(plugin => {
+    pageConfig.usingComponents[plugin] = usingPlugins[plugin].path;
+  });
 
   addFileToCompilation(compilation, {
     filename: `${pageRoute}.json`,