+++ packages/rax-miniapp-runtime-webpack-plugin/src/generators/root.js
@@ -1,28 +1,24 @@
-const adapter = require('../adapter');
-const ejs = require('ejs');
+const { platformMap } = require('miniapp-builder-shared');
+
 const addFileToCompilation = require('../utils/addFileToCompilation');
-const getTemplate = require('../utils/getTemplate');
+const { buildTemplate, buildSjs, buildNativeComponentTemplate } = require('./templates');
 
 function generateRootTmpl(
   compilation,
-  { usingPlugins, usingComponents, target, command, pluginDir }
+  { usingPlugins, usingComponents, target, command, modifyTemplate }
 ) {
-  const template = ejs.render(getTemplate(pluginDir, 'root.xml', target));
-  const pluginTmpl = ejs.render(getTemplate(pluginDir, 'plugin.xml', target), {
-    usingPlugins
-  });
-  const componentTmpl = ejs.render(getTemplate(pluginDir, 'custom-component.xml', target), {
-    usingComponents
+  const template = buildTemplate(target, modifyTemplate);
+  const sjs = buildSjs(target);
+  const nativeComponentTemplate = buildNativeComponentTemplate(usingPlugins, target) + buildNativeComponentTemplate(usingComponents, target);
+
+  addFileToCompilation(compilation, {
+    filename: `root${platformMap[target].extension.xml}`,
+    content: template + nativeComponentTemplate,
+    target, command
   });
   addFileToCompilation(compilation, {
-    filename: `root.${adapter[target].xml}`,
-    content: template + pluginTmpl + componentTmpl,
-    target,
-    command,
-  });
-  addFileToCompilation(compilation, {
-    filename: `tool.${adapter[target].script}`,
-    content: ejs.render(getTemplate(pluginDir, `tool.${adapter[target].script}`, target)),
+    filename: `tool${platformMap[target].extension.script}`,
+    content: sjs,
     target,
     command,
   });