+++ packages/rax-miniapp-runtime-webpack-plugin/src/platforms/index.js
@@ -1,0 +1,8 @@
+const ali = require('./ali');
+const wechat = require('./wechat');
+const { constants: { MINIAPP, WECHAT_MINIPROGRAM }} = require('miniapp-builder-shared');
+
+module.exports = {
+  [MINIAPP]: ali,
+  [WECHAT_MINIPROGRAM]: wechat
+};