+++ packages/rax-miniapp-runtime-webpack-plugin/src/utils/wrapChunks.js
@@ -1,10 +1,10 @@
 const ModuleFilenameHelpers = require('webpack/lib/ModuleFilenameHelpers');
 const { RawSource, ConcatSource } = require('webpack-sources');
+const { platformMap } = require('miniapp-builder-shared');
 const { readFileSync } = require('fs-extra');
 const { resolve } = require('path');
 const adjustCSS = require('../utils/adjustCSS');
-const adapter = require('../adapter');
-const { WECHAT_MINIPROGRAM } = require('../constants');
+const { UNRECURSIVE_TEMPLATE_TYPE } = require('../constants');
 
 const matchFile = (fileName, ext) =>
   ModuleFilenameHelpers.matchObject(
@@ -12,19 +12,17 @@ const matchFile = (fileName, ext) =>
     fileName
   );
 
+const FunctionPolyfill = 'Function||(Function=function(){return function(){return Symbol}}),void 0===Function.prototype.call&&(Function.prototype.call=function(n){(n=n||window).fn=this;const t=[...arguments].slice(1),o=n.fn(...t);return delete n.fn,o}),void 0===Function.prototype.apply&&(Function.prototype.apply=function(n){let t;return(n=n||window).fn=this,t=arguments[1]?n.fn(...arguments[1]):n.fn(),delete n.fn,t})';
+
 // Add content to chunks head and tail
 module.exports = function(compilation, chunks, pluginDir, target) {
-  const FunctionPolyfill = readFileSync(
-    resolve(pluginDir, 'templates', 'FunctionPolyfill.js.ejs'),
-    'utf-8'
-  );
   chunks.forEach((chunk) => {
     chunk.files.forEach((fileName) => {
       if (matchFile(fileName, 'js')) {
         // Page js
         const headerContent =
-          `${FunctionPolyfill}
-          module.exports = function(window, document) {const HTMLElement = window["HTMLElement"];`;
+`${FunctionPolyfill}
+module.exports = function(window, document) {const HTMLElement = window["HTMLElement"];`;
 
         const footerContent = '}';
 
@@ -35,9 +33,9 @@ module.exports = function(compilation, chunks, pluginDir, target) {
         );
       } else if (matchFile(fileName, 'css')) {
         compilation.assets[
-          `${fileName}.${adapter[target].css}`
+          `${fileName}${platformMap[target].extension.css}`
         ] = new RawSource(
-          adjustCSS(compilation.assets[fileName].source(), target === WECHAT_MINIPROGRAM)
+          adjustCSS(compilation.assets[fileName].source(), UNRECURSIVE_TEMPLATE_TYPE.has(target))
         );
       }
     });