+++ packages/rax-miniapp-config-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.1] - 2021-03-11
+
+### Changed
+
+- Use constants from miniapp-builder-shared
+
 ## [2.0.1] - 2021-02-25
 
 ### Added